+++ packages/charts/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @react-magma/charts
 
+## 14.0.0-rc.3
+
+### Patch Changes
+
+- cb81479: chore: add optional `ariaLabel` property for SVG container
+
+## 14.0.0-rc.2
+
+### Patch Changes
+
+- 7c8c1ea: fix(Charts): Fix magma styles not applied to carbon charts.
+
 ## 14.0.0-rc.1
 
 ### Patch Changes